refactor: simplify build and packaging of input-group components
