+++ packages/components/input-group/css/bs-input-group-prepend.css.js
