+++ packages/components/input-group/css/bs-input-group-small.css.js
