+++ packages/components/input-group/css/index.js
@@ -1,10 +1,6 @@
-export { BsInputGroup } from './bs-input-group.js';
 export { BsInputGroupCss } from './bs-input-group.css.js';
-export { BsInputGroupText } from './bs-input-group-text.js';
 export { BsInputGroupTextCss } from './bs-input-group-text.css.js';
-export { BsInputGroupAppend } from './bs-input-group-append.js';
 export { BsInputGroupAppendCss } from './bs-input-group-append.css.js';
-export { BsInputGroupPrepend } from './bs-input-group-prepend.js';
 export { BsInputGroupPrependCss } from './bs-input-group-prepend.css.js';
 export { BsInputGroupLargeCss } from './bs-input-group-large.css.js';
 export { BsInputGroupSmallCss } from './bs-input-group-small.css.js';
