+++ packages/components/input-group/index.js
@@ -1,0 +1,5 @@
+export * from './css'
+export { BsInputGroup } from './bs-input-group.js';
+export { BsInputGroupText } from './bs-input-group-text.js';
+export { BsInputGroupAppend } from './bs-input-group-append.js';
+export { BsInputGroupPrepend } from './bs-input-group-prepend.js';
